+++ src/App/Core/ViewModels/LayerListViewModel.cs
@@ -353,7 +353,7 @@ namespace WhereToFly.App.Core.ViewModels
 
             string webSiteToOpen = this.downloadWebSiteList[result];
 
-            await Xamarin.Essentials.Launcher.OpenAsync(webSiteToOpen);
+            await Browser.OpenAsync(webSiteToOpen, BrowserLaunchMode.External);
         }
 
         /// <summary>