+++ Changelog.md
@@ -16,6 +16,7 @@ Released on ?.
 
 - Fixed loading weather icons for some pages
 - Fixed map starting location, using last zoomed-to location
+- Fixed opening websites in external browser window
 
 ## Version 1.11.0